:+1:エラーの関数をまとめた
