+++ src/builtin_func/builtin_exit.c
@@ -2,6 +2,7 @@
 #include "../../header/standard.h"
 #include "../../libft/libft.h"
 #include "../../header/builtin_func.h"
+#include "../../header/print.h"
 
 long	ft_atoll(const char *str)
 {
